file: employeeModel.ts
