+++ src/api/v1/models/employeeModel.ts
@@ -1,0 +1,13 @@
+/**
+ * @interface Employee
+ * @description Represents an employee object
+ */
+export type Employee = {
+  id: string;
+  name: string;
+  position: string;
+  department: string;
+  email: string;
+  phone: string;
+  branch: string;
+};
