+++ test/validate.test.ts
@@ -1,7 +1,7 @@
 import { validate } from "../src/api/v1/middleware/validate";
 import { employeeSchema } from "../src/api/v1/validation/employeeValidation";
 import { branchSchema } from "../src/api/v1/validation/branchValidation";
-import { Employee } from "src/api/v1/services/employeeServices";
+import { Employee } from "src/api/v1/models/employeeModel";
 import { Branch } from "src/api/v1/models/branchModel";
 
 describe("Validate function for employees", () => {
